agent-project/2.ai-cooking/requirements.txt
streamlit
openai
agno
python-dotenv
duckduckgo-search
lancedb
tantivy
pypdf
sqlalchemy
yfinance
google-genai
groq
anthropic
huggingface_hub
ollama
